# ci/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "VideoOrganizer API",
    "description": "Self-hosted media library manager \u2014 videos, tags, custom properties, background workers (thumbnail sprites \u002B Md5 dedup), and bulk import.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/videos": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "CreateVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Video"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos/{id}": {
      "get": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "GetVideoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "UpdateVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "VideoOrganizer.API"
        ],
        "operationId": "DeleteVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/backup/settings": {
      "get": {
        "tags": [
          "Backup"
        ],
        "operationId": "GetBackupSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Backup"
        ],
        "operationId": "SetBackupDirectory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBackupDirectoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupSettings"
                }
              }
            }
          }
        }
      }
    },
    "/api/backup/{fileName}/download": {
      "get": {
        "tags": [
          "Backup"
        ],
        "operationId": "DownloadBackup",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/backup/{fileName}": {
      "delete": {
        "tags": [
          "Backup"
        ],
        "operationId": "DeleteBackup",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/backup/{fileName}/restore": {
      "post": {
        "tags": [
          "Backup"
        ],
        "operationId": "RestoreBackup",
        "parameters": [
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupSettings": {
        "required": [
          "directory",
          "writable",
          "error"
        ],
        "type": "object",
        "properties": {
          "directory": {
            "type": "string"
          },
          "writable": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CameraTypes": {
        "enum": [
          "unknown",
          "cellPhone",
          "hiddenCamera",
          "camcorder",
          "professionalCamera",
          "notChecked"
        ],
        "type": "string"
      },
      "ChapterMarker": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "ChapterMarkerDto": {
        "required": [
          "offset",
          "comment"
        ],
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DomainCameraTypes": {
        "enum": [
          "unknown",
          "cellPhone",
          "hiddenCamera",
          "camcorder",
          "professionalCamera",
          "notChecked"
        ],
        "type": "string"
      },
...
      "DomainVideoCodec": {
        "enum": [
          "h264",
          "h265",
          "hevc",
          "notChecked",
          "other"
        ],
        "type": "string"
      },
      "DomainVideoDimensionFormat": {
        "enum": [
          "uhd8k",
          "uhd4K",
          "hd1080p",
          "hd720p",
          "sd576p4x3",
          "sd576p16x9",
          "sd480p4x3",
          "sd480p16x9",
          "verticalUHD4k",
          "vertical1080p",
          "vertical720p",
          "nonStandard"
        ],
        "type": "string"
      },
      "DomainVideoQuality": {
        "enum": [
          "singleCamera",
          "multipleCameras",
          "lowQuality",
          "notChecked"
        ],
        "type": "string"
      },
...
      "OcrTextLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Video"
              }
            ]
          },
          "timeSeconds": {
            "type": "number",
            "format": "double"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "PropertyValueDto": {
        "required": [
          "propertyDefinitionId",
          "propertyName",
          "dataType",
          "value"
        ],
        "type": "object",
        "properties": {
          "propertyDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "propertyName": {
            "type": "string"
          },
          "dataType": {
            "$ref": "#/components/schemas/PropertyDataTypeDto"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "SetBackupDirectoryRequest": {
        "required": [
          "directory"
        ],
        "type": "object",
        "properties": {
          "directory": {
            "type": "string"
          }
        }
      },
...
      "UpdateVideoRequest": {
        "required": [
          "fileName",
          "ingestDate",
          "cameraType",
          "videoQuality",
          "watchCount",
          "notes",
          "needsReview",
          "isFavorite",
          "clipStartSeconds",
          "clipEndSeconds",
          "chapterMarkers",
          "videoBlocks",
          "tagIds",
          "properties"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/CameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/VideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chapterMarkers": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ChapterMarkerDto"
            }
          },
          "videoBlocks": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/VideoBlockDto"
            }
          },
          "tagIds": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "properties": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/PropertyValueWrite"
            }
          }
        }
      },
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/DomainVideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/DomainVideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/DomainCameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/DomainVideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "clipExported": {
            "type": "boolean"
          },
          "exportedToVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "ocrScannedThroughSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarker"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlock"
            }
          },
          "ocrTextLines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcrTextLine"
            }
          },
          "videoTags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTag"
            }
          },
          "propertyValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoPropertyValue"
            }
          }
        }
      },
      "VideoBlock": {
        "type": "object",
        "properties": {
          "offsetInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "lengthInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "videoBlockType": {
            "$ref": "#/components/schemas/DomainVideoBlockTypes"
          }
        }
      },
      "VideoBlockDto": {
        "required": [
          "offsetInSeconds",
          "lengthInSeconds",
          "videoBlockType"
        ],
        "type": "object",
        "properties": {
          "offsetInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "lengthInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "videoBlockType": {
            "$ref": "#/components/schemas/VideoBlockTypes"
          }
        }
      },
      "VideoBlockTypes": {
        "enum": [
          "clip",
          "hide",
          "other"
        ],
        "type": "string"
      },
      "VideoCodec": {
        "enum": [
          "h264",
          "h265",
          "hevc",
          "notChecked",
          "other"
        ],
        "type": "string"
      },
      "VideoDimensionFormat": {
        "enum": [
          "uhd8k",
          "uhd4K",
          "hd1080p",
          "hd720p",
          "sd576p4x3",
          "sd576p16x9",
          "sd480p4x3",
          "sd480p16x9",
          "verticalUHD4k",
          "vertical1080p",
          "vertical720p",
          "nonStandard"
        ],
        "type": "string"
      },
      "VideoDto": {
        "required": [
          "id",
          "fileName",
          "filePath",
          "md5",
          "md5Failed",
          "md5FailedError",
          "thumbnailsFailed",
          "thumbnailsFailedError",
          "thumbnailsGenerated",
          "importJobId",
          "fileSize",
          "duration",
          "height",
          "width",
          "videoDimensionFormat",
          "videoCodec",
          "bitrate",
          "frameRate",
          "pixelFormat",
          "ratio",
          "creationTime",
          "videoStreamCount",
          "audioStreamCount",
          "ingestDate",
          "cameraType",
          "videoQuality",
          "watchCount",
          "notes",
          "needsReview",
          "playbackIssue",
          "markedForDeletion",
          "isFavorite",
          "parentVideoId",
          "clipStartSeconds",
          "clipEndSeconds",
          "isClip",
          "isExportedClip",
          "isEdited",
          "chapterMarkers",
          "videoBlocks",
          "tags",
          "properties"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "md5": {
            "type": [
              "null",
              "string"
            ]
          },
          "md5Failed": {
            "type": "boolean"
          },
          "md5FailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsFailed": {
            "type": "boolean"
          },
          "thumbnailsFailedError": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumbnailsGenerated": {
            "type": "boolean"
          },
          "importJobId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "duration": {
            "pattern": "^-?(\\d\u002B\\.)?\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,7})?$",
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "videoDimensionFormat": {
            "$ref": "#/components/schemas/VideoDimensionFormat"
          },
          "videoCodec": {
            "$ref": "#/components/schemas/VideoCodec"
          },
          "bitrate": {
            "type": "integer",
            "format": "int64"
          },
          "frameRate": {
            "type": "number",
            "format": "double"
          },
          "pixelFormat": {
            "type": [
              "null",
              "string"
            ]
          },
          "ratio": {
            "type": [
              "null",
              "string"
            ]
          },
          "creationTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "videoStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "ingestDate": {
            "type": "string",
            "format": "date-time"
          },
          "cameraType": {
            "$ref": "#/components/schemas/CameraTypes"
          },
          "videoQuality": {
            "$ref": "#/components/schemas/VideoQuality"
          },
          "watchCount": {
            "type": "integer",
            "format": "int32"
          },
          "notes": {
            "type": "string"
          },
          "needsReview": {
            "type": "boolean"
          },
          "playbackIssue": {
            "type": "boolean"
          },
          "markedForDeletion": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentVideoId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clipStartSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "clipEndSeconds": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isClip": {
            "type": "boolean"
          },
          "isExportedClip": {
            "type": "boolean"
          },
          "isEdited": {
            "type": "boolean"
          },
          "chapterMarkers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterMarkerDto"
            }
          },
          "videoBlocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoBlockDto"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoTagSummaryDto"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyValueDto"
            }
          }
        }
      },
      "VideoPropertyValue": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Video"
              }
            ]
          },
          "propertyDefinitionId": {
            "type": "string",
            "format": "uuid"
          },
          "propertyDefinition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PropertyDefinition"
              }
            ]
          },
          "value": {
            "type": "string"
          }
        }
      },
      "VideoQuality": {
        "enum": [
          "singleCamera",
          "multipleCameras",
          "lowQuality",
          "notChecked"
        ],
        "type": "string"
      },
...
      "VideoTag": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "string",
            "format": "uuid"
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Video"
              }
            ]
          },
          "tagId": {
            "type": "string",
            "format": "uuid"
          },
          "tag": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Tag"
              }
            ]
          }
        }
      },
      "VideoTagSummaryDto": {
        "required": [
          "id",
          "tagGroupId",
          "tagGroupName",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tagGroupId": {
            "type": "string",
            "format": "uuid"
          },
          "tagGroupName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}